pyproject.toml
[project]
name = "project-itnewsapp-line-notice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "googletrans==3.1.0a0",
    "line-bot-sdk>=3.21.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.14.10",
    "timedelta>=2020.12.3",
]
